Rename sql to postgres
