+++ postgres/eventstore/strategy/single_table_strategy.go
