+++ postgres/eventstore/strategy/stream_iterator.go
